Replaced the Aadl2OutlineTreeProvider with Aadl2BackgroundOutlineTreeProvider.  Got rid of Aadl2BackgroundOutlineTreeProvider.  Original "non-background" version persists in the history of the file.  Same of the error model outline.
