+++ core/org.osate.xtext.aadl2.ui/src/org/osate/xtext/aadl2/ui/Aadl2UiModule.java
@@ -38,8 +38,6 @@ import org.eclipse.xtext.ui.editor.hover.IEObjectHoverProvider;
 import org.eclipse.xtext.ui.editor.model.IResourceForEditorInputFactory;
 import org.eclipse.xtext.ui.editor.model.XtextDocumentProvider;
 import org.eclipse.xtext.ui.editor.model.edit.ITextEditComposer;
-import org.eclipse.xtext.ui.editor.outline.IOutlineTreeProvider;
-import org.eclipse.xtext.ui.editor.outline.impl.IOutlineTreeStructureProvider;
 import org.eclipse.xtext.ui.refactoring.IDependentElementsCalculator;
 import org.eclipse.xtext.ui.resource.IStorage2UriMapper;
 import org.eclipse.xtext.ui.shared.Access;
@@ -53,7 +51,6 @@ import org.osate.xtext.aadl2.ui.editor.model.Aadl2DocumentProvider;
 import org.osate.xtext.aadl2.ui.editor.model.Aadl2ResourceForEditorInputFactory;
 import org.osate.xtext.aadl2.ui.editor.model.edit.FormattingTextEditComposer;
 import org.osate.xtext.aadl2.ui.editor.occurrences.Aadl2OccurrenceComputer;
-import org.osate.xtext.aadl2.ui.outline.Aadl2BackgroundOutlineTreeProvider;
 import org.osate.xtext.aadl2.ui.outline.Aadl2OutlinePage;
 import org.osate.xtext.aadl2.ui.refactoring.Aadl2DependentElementsCalculator;
 import org.osate.xtext.aadl2.ui.refactoring.impl.Aadl2RenameStrategy;
@@ -165,15 +162,5 @@ public class Aadl2UiModule extends org.osate.xtext.aadl2.ui.AbstractAadl2UiModul
 	 */
 	public Class<? extends IEObjectHoverProvider> bindIEObjectHoverProvider() {
 		return Aadl2EObjectHoverProvider.class;
-	}
-
-	@Override
-	public Class<? extends IOutlineTreeProvider> bindIOutlineTreeProvider() {
-		return Aadl2BackgroundOutlineTreeProvider.class;
-	}
-
-	@Override
-	public Class<? extends IOutlineTreeStructureProvider> bindIOutlineTreeStructureProvider() {
-		return Aadl2BackgroundOutlineTreeProvider.class;
 	}
 }
